Use adn-timeline from @imathis.
This is so much better than what I was using (tweets posing as app.net
posts).  I would have liked to use the new markup but alas, I'm not a
designer.  It was easier for me to use a custom render function and map
it to the markup that was in place and already styled.  Yes it's a hack
but it works for now.